// src/client/zod.gen.ts
export const zPersonalProfileCreate = z.object({
    korean_name: z.string(),
    english_name: z.string(),
    birth_date: z.iso.date(),
    chinese_name: z.optional(z.union([
        z.string(),
        z.null()
    ])),
    resident_number: z.optional(z.union([
        z.string(),
        z.null()
    ])),
    blessing_date: z.optional(z.union([
        z.iso.date(),
        z.null()
    ])),
    appointment_date: z.optional(z.union([
        z.iso.date(),
        z.null()
    ])),
    nationality: z.optional(z.union([
        z.string(),
        z.null()
    ])).default('대한민국'),
    organization: z.optional(z.union([
        z.string(),
        z.null()
    ])),
    department: z.optional(z.union([
        z.string(),
        z.null()
    ])),
    position: z.optional(z.union([
        z.string(),
        z.null()
    ])),
    mobile_phone: z.optional(z.union([
        z.string(),
        z.null()
    ])),
    extension: z.optional(z.union([
        z.string(),
        z.null()
    ])),
    email: z.optional(z.union([
        z.string(),
        z.null()
    ])),
    address: z.optional(z.union([
        z.string(),
        z.null()
    ])),
    remarks: z.optional(z.union([
        z.string(),
        z.null()
    ])),
    profile_update_date: z.optional(z.union([
        z.iso.date(),
        z.null()
    ])),
    resume_id: z.optional(z.union([
        z.uuid(),
        z.null()
    ]))
});